refactor!: move chain definitions out of aa-sdk core
