+++ aa-sdk/core/src/client/smartAccountClient.test.ts
@@ -1,5 +1,5 @@
-import { polygonMumbai } from "@aa-sdk/core";
 import { custom, type Transaction } from "viem";
+import { polygonMumbai } from "viem/chains";
 import type { SpyInstance } from "vitest";
 import * as receiptActions from "../actions/bundler/getUserOperationReceipt.js";
 import type { UserOperationReceipt } from "../types.js";
