+++ account-kit/smart-contracts/plugindefs/multisig/config.ts
@@ -13,7 +13,7 @@ import {
   sepolia,
   zora,
   zoraSepolia,
-} from "@aa-sdk/core";
+} from "@account-kit/infra";
 import type { PluginConfig } from "@account-kit/plugingen";
 import { parseAbiParameters } from "viem";
 import { MultisigPluginAbi } from "./abi.js";
