+++ account-kit/smart-contracts/src/msca/utils.ts
@@ -1,6 +1,14 @@
 import {
   AccountNotFoundError,
   ChainNotFoundError,
+  type GetAccountParameter,
+  type SmartAccountClient,
+  type SmartAccountSigner,
+  type SmartContractAccount,
+  type SmartContractAccountWithSigner,
+  type UpgradeToData,
+} from "@aa-sdk/core";
+import {
   arbitrum,
   arbitrumSepolia,
   base,
@@ -11,13 +19,7 @@ import {
   polygon,
   polygonAmoy,
   sepolia,
-  type GetAccountParameter,
-  type SmartAccountClient,
-  type SmartAccountSigner,
-  type SmartContractAccount,
-  type SmartContractAccountWithSigner,
-  type UpgradeToData,
-} from "@aa-sdk/core";
+} from "@account-kit/infra";
 import type { Address, Chain, Transport } from "viem";
 import {
   custom,
